Move wrapPromise to helpers/promise
