+++ index.js
@@ -1,6 +1,7 @@
 const patch = require('./lib/patch')
 const tracer = require('./lib/core/tracer')
 const util = require('util')
+const promises = require('./lib/helpers/promises')
 
 patch()
 
@@ -8,7 +9,7 @@ module.exports = {
   configure: tracer.configure,
   currentSpan: tracer.currentSpan,
   startSpan: tracer.startSpan,
-  wrapPromise: tracer.wrapPromise,
+  wrapPromise: promises.wrapPromise,
 
   /**
     * @deprecated since version 1.1.0
